Created LogEvents class
